backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic<2.0.0,>=1.7.2
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic>=1.11.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
requests==2.31.0
fhir.resources==7.0.2
redis==5.0.1
